requirements.txt
absl-py==0.15.0
appdirs==1.4.4
argcomplete==1.12.3
astunparse==1.6.3
attrs==21.2.0
audioread==2.1.9
bleach==4.1.0
build==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
check-manifest==0.46
clang==5.0
click==8.0.3
click-log==0.3.2
colorama==0.4.4
commitizen==2.17.13
cryptography==36.0.0
cycler==0.11.0
dataclasses-json==0.5.6
decli==0.5.2
decorator==5.1.0
docutils==0.18.1
dotty-dict==1.3.0
ffpyplayer==4.3.2
flake8==3.9.2
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.24
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
invoke==1.6.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
keyring==23.4.0
kiwisolver==1.3.2
librosa==0.8.1
llvmlite==0.34.0
Markdown==3.3.4
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-enum==1.5.1
matplotlib==3.4.3
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
numba==0.51.2
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.2
pandas==1.3.0
pep517==0.12.0
Pillow==8.4.0
pkginfo==1.8.2
pluggy==0.13.1
pooch==1.5.2
prompt-toolkit==3.0.23
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycodestyle==2.7.0
pycparser==2.21
pydub==0.24.1
pyflakes==2.3.1
Pygments==2.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
python-gitlab==2.10.1
python-semantic-release==7.17.0
pytorchcv==0.0.66
pytz==2021.3
PyYAML==5.4.1
questionary==1.10.0
readme-renderer==30.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
resampy==0.2.2
rfc3986==1.5.0
rsa==4.7.2
scikit-learn==0.24.0
scipy==1.7.2
SecretStorage==3.3.1
semver==2.13.0
setuptools-scm==6.3.2
six==1.15.0
smmap==5.0.0
SoundFile==0.10.3.post1
stringcase==1.2.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
tomlkit==0.7.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.2
twine==3.7.0
typed-ast==1.4.3
types-PyYAML==5.4.6
typing-extensions==3.7.4
typing-inspect==0.7.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
